refactor: split up logic of defaulting & parsing
